Add files via upload
This .py can produce a random wallet with 12 mnemonic phrases.  Maybe someone would hit 1 address with Ethereum or other tokens in there some day.
*Reminder:
The number of 0x.... addresses with Ethereum in it now is about ~2 M, or 2E6, and the theoretical address produced from 12 mnemonic phrases is about 1E38.
Good Luck!